Refactored ActiveRecordChangeTracking.

+++ lib/ruby_sync/connectors/active_record_change_tracking.rb
@@ -26,14 +26,14 @@ module RubySync::Connectors::ActiveRecordChangeTracking
   def update_mirror(path); end
   def delete_from_mirror(path); end
 
-  def each_change
+  def each_change(&blk)
     # Process each RubySyncEvent and then delete it from the db.
-    if Object.const_defined?(:RubySyncEvent) and @models.include?('RubySyncEvent')
+    if Object.const_defined?(:RubySyncEvent) and @models.include?('RubySyncEvent')#TODO used a constant instead of a hard coded model name
       ::RubySyncEvent.find(:all).each do |rse|
         event = RubySync::Event.new(rse.event_type, self, rse.trackable_id, nil, to_payload(rse))
         yield event
 
-        #usefull ?
+        #useful ?
         if is_vault? and @pipeline and rse.event_type==:delete
           association = association_for @pipeline.association_context, rse.trackable_id
           remove_association association
@@ -41,36 +41,42 @@ module RubySync::Connectors::ActiveRecordChangeTracking
 
         ::RubySyncEvent.delete rse
       end
-    else
-      # scan existing entries to see if any new or modified
-      each_entry.each do |entry|
-        digest = digest(entry)
-        unless stored_digest = track_class.find_by_key(entry.send(:"#{ar_class.primary_key}")) and digest == stored_digest
-          operations = create_operations_for(entry_from_active_record(entry))
-          yield RubySync::Event.add(self, entry.send(:"#{ar_class.primary_key}"), nil, operations)
-          track_class.create(:key => entry.send(:"#{ar_class.primary_key}"), :digest => digest)
-        end
-      end
-
-      # scan track to find deleted
-      track_class.find(:all).each do |record|
-        key=record.key
-        unless self[key]
-          yield RubySync::Event.delete(self, key)
-          track_class.destroy_all(:key => key)
-          if is_vault? and @pipeline
-            association = association_for @pipeline.association_context, key
-            remove_association association
+    elsif respond_to? :track
+      if !track_class.nil?
+        # scan existing entries to see if any new or modified
+        each_entry do |path, entry|
+          digest = digest(entry)#TODO used entry.hash instead of digest ?
+          unless stored_digest = track_class.find_by_key(path) and digest == stored_digest
+            operations = create_operations_for(entry)
+            yield RubySync::Event.add(self, path, nil, operations)
+            track_class.create(:key => path, :digest => digest)
           end
         end
+        
+        # scan track to find deleted entries
+        track_class.find(:all).each do |record|
+          key=record.key
+          unless self[key]
+            yield RubySync::Event.delete(self, key)
+            track_class.destroy_all(:key => key)
+            if is_vault? and @pipeline
+              association = association_for @pipeline.association_context, key
+              remove_association association
+            end
+          end
+        end
+      elsif track.is_a?(RubySync::Connectors::LdapChangelogRubyConnector)
+        log.debug 'Delegate #each_change to LDAP Changelog system'
+        track.each_change(&blk)# delegate the change tracking to LDAP Changelog system
       end
-    end
-    
+    else
+      super
+    end  
 
   end
 
   def digest(o)
-    Digest::MD5.hexdigest(entry_from_active_record(o).to_yaml)
+    Digest::MD5.hexdigest(o.to_yaml)
   end
 
   # Create a hash suitable to use as rubysync event payload
